refactor: 데이터 받아온 후 가공하는 로직으로 분리
